+++ pycm/__init__.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 """PyCM modules."""
-from .pycm_param import PYCM_VERSION, OVERALL_BENCHMARK_LIST, CLASS_BENCHMARK_LIST
+from .params import PYCM_VERSION, OVERALL_BENCHMARK_LIST, CLASS_BENCHMARK_LIST
 from .pycm_error import *
 from .pycm_output import pycm_help, online_help
 from .pycm_distance import DistanceType