+++ pycm/pycm_class_func.py
@@ -4,7 +4,7 @@ from __future__ import division
 import math
 from .utils import normal_quantile
 from .pycm_interpret import *
-from .pycm_param import CLASS_PARAMS
+from .params import CLASS_PARAMS
 
 
 def sensitivity_index_calc(TPR, FPR):