+++ pycm/pycm_output.py
@@ -2,7 +2,7 @@
 """Outputs functions."""
 from __future__ import division
 from functools import partial
-from .pycm_param import *
+from .params import *
 from .utils import rounder, sort_char_num
 import webbrowser